+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/tests/latest/test_network_commands.py
@@ -1963,12 +1963,9 @@ class NetworkActiveActiveVnetScenarioTest(ScenarioTest):  # pylint: disable=too-
         self.cmd('network vnet-gateway wait -g {rg} -n {gw1} --created')
         self.cmd('network vnet-gateway wait -g {rg} -n {gw2} --created')
 
-        # TODO: Re-enabled once issue https://github.com/Azure/azure-cli/issues/7977 is resolved
         # create and connect the VNet gateways
-        with self.assertRaisesRegexp(CLIError, '255.255.255.255'):
-            self.cmd('network vpn-connection create -g {rg} -n {conn12} --vnet-gateway1 {gw1} --vnet-gateway2 {gw2} --shared-key {key} --enable-bgp')
-        with self.assertRaisesRegexp(CLIError, '255.255.255.255'):
-            self.cmd('network vpn-connection create -g {rg} -n {conn21} --vnet-gateway1 {gw2} --vnet-gateway2 {gw1} --shared-key {key} --enable-bgp')
+        self.cmd('network vpn-connection create -g {rg} -n {conn12} --vnet-gateway1 {gw1} --vnet-gateway2 {gw2} --shared-key {key} --enable-bgp')
+        self.cmd('network vpn-connection create -g {rg} -n {conn21} --vnet-gateway1 {gw2} --vnet-gateway2 {gw1} --shared-key {key} --enable-bgp')
 
 
 class NetworkVpnGatewayScenarioTest(ScenarioTest):
@@ -2110,6 +2107,48 @@ class NetworkTrafficManagerScenarioTest(ScenarioTest):
                  checks=self.check('length(@)', 0))
 
         self.cmd('network traffic-manager profile delete -g {rg} -n {tm}')
+
+    @ResourceGroupPreparer('cli_test_traffic_manager_subnet')
+    def test_network_traffic_manager_subnet_routing(self, resource_group):
+
+        self.kwargs.update({
+            'tm': 'tm1',
+            'endpoint': 'ep1',
+            'dns': self.create_random_name('testtm', 20),
+            'pip': 'ip1',
+            'ip_dns': self.create_random_name('testpip', 20)
+        })
+
+        self.cmd('network traffic-manager profile create -n {tm} -g {rg} --routing-method subnet --unique-dns-name {dns} --custom-headers foo=bar --status-code-ranges 200-202', checks=[
+            self.check('TrafficManagerProfile.monitorConfig.expectedStatusCodeRanges[0].min', 200),
+            self.check('TrafficManagerProfile.monitorConfig.expectedStatusCodeRanges[0].max', 202),
+            self.check('TrafficManagerProfile.monitorConfig.customHeaders[0].name', 'foo'),
+            self.check('TrafficManagerProfile.monitorConfig.customHeaders[0].value', 'bar')
+        ])
+        self.kwargs['ip_id'] = self.cmd('network public-ip create -g {rg} -n {pip} --dns-name {ip_dns} --query publicIp.id').get_output_in_json()
+        self.cmd('network traffic-manager profile update -n {tm} -g {rg} --status-code-ranges 200-204 --custom-headers foo=doo test=best', checks=[
+            self.check('monitorConfig.expectedStatusCodeRanges[0].min', 200),
+            self.check('monitorConfig.expectedStatusCodeRanges[0].max', 204),
+            self.check('monitorConfig.customHeaders[0].name', 'foo'),
+            self.check('monitorConfig.customHeaders[0].value', 'doo'),
+            self.check('monitorConfig.customHeaders[1].name', 'test'),
+            self.check('monitorConfig.customHeaders[1].value', 'best')
+        ])
+
+        # Endpoint tests
+        self.cmd('network traffic-manager endpoint create -n {endpoint} --profile-name {tm} -g {rg} --type azureEndpoints --target-resource-id {ip_id} --subnets 10.0.0.0 --custom-headers test=best', checks=[
+            self.check('customHeaders[0].name', 'test'),
+            self.check('customHeaders[0].value', 'best'),
+            self.check('subnets[0].first', '10.0.0.0')
+        ])
+        self.cmd('network traffic-manager endpoint update -n {endpoint} --type azureEndpoints --profile-name {tm} -g {rg} --subnets 10.0.0.0:24', checks=[
+            self.check('subnets[0].first', '10.0.0.0'),
+            self.check('subnets[0].scope', '24')
+        ])
+        self.cmd('network traffic-manager endpoint update -n {endpoint} --type azureEndpoints --profile-name {tm} -g {rg} --subnets 10.0.0.0-11.0.0.0', checks=[
+            self.check('subnets[0].first', '10.0.0.0'),
+            self.check('subnets[0].last', '11.0.0.0')
+        ])
 
 
 class NetworkWatcherConfigureScenarioTest(LiveScenarioTest):