+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_help.py
@@ -546,48 +546,81 @@ helps['vm diagnostics set'] = """
     short-summary: Configure the Azure VM diagnostics extension.
     examples:
         - name: Set up default diagnostics on a Linux VM for Azure Portal VM metrics graphs and syslog collection.
-          text: >
-            # Set the following 3 parameters first.\n
-            my_resource_group=<Resource group name containing your Linux VM and the storage account>\n
-            my_linux_vm=<Your Azure Linux VM name>\n
-            my_diagnostic_storage_account=<Your Azure storage account for storing VM diagnostic data>\n
+          text: |
+                # Set the following 3 parameters first.
+                my_resource_group=<Resource group name containing your Linux VM and the storage account>
+                my_linux_vm=<Your Azure Linux VM name>
+                my_diagnostic_storage_account=<Your Azure storage account for storing VM diagnostic data>
 
-            my_vm_resource_id=$(az vm show -g $my_resource_group -n $my_linux_vm --query "id" -o tsv)\n
+                my_vm_resource_id=$(az vm show -g $my_resource_group -n $my_linux_vm --query "id" -o tsv)
 
-            default_config=$(az vm diagnostics get-default-config \\
-                | sed "s#__DIAGNOSTIC_STORAGE_ACCOUNT__#$my_diagnostic_storage_account#g" \\
-                | sed "s#__VM_OR_VMSS_RESOURCE_ID__#$my_vm_resource_id#g")
+                default_config=$(az vm diagnostics get-default-config \\
+                    | sed "s#__DIAGNOSTIC_STORAGE_ACCOUNT__#$my_diagnostic_storage_account#g" \\
+                    | sed "s#__VM_OR_VMSS_RESOURCE_ID__#$my_vm_resource_id#g")
 
-            storage_sastoken=$(az storage account generate-sas \\
-                --account-name $my_diagnostic_storage_account --expiry 2037-12-31T23:59:00Z \\
-                --permissions wlacu --resource-types co --services bt -o tsv)
+                storage_sastoken=$(az storage account generate-sas \\
+                    --account-name $my_diagnostic_storage_account --expiry 2037-12-31T23:59:00Z \\
+                    --permissions wlacu --resource-types co --services bt -o tsv)
 
-            protected_settings="{'storageAccountName': '{my_diagnostic_storage_account}', \\
-                'storageAccountSasToken': '{storage_sastoken}'}"
+                protected_settings="{'storageAccountName': '$my_diagnostic_storage_account', \\
+                    'storageAccountSasToken': '$storage_sastoken'}"
 
-            az vm diagnostics set --settings "{default_config}" \\
-                --protected-settings "{protected_settings}" \\
-                --resource-group $my_resource_group --vm-name $my_linux_vm
+                az vm diagnostics set --settings "$default_config" \\
+                    --protected-settings "$protected_settings" \\
+                    --resource-group $my_resource_group --vm-name $my_linux_vm
+
+        - name: Set up default diagnostics on a Windows VM.
+          text: |
+                # Set the following 3 parameters first.
+                my_resource_group=<Resource group name containing your Windows VM and the storage account>
+                my_windows_vm=<Your Azure Windows VM name>
+                my_diagnostic_storage_account=<Your Azure storage account for storing VM diagnostic data>
+
+                my_vm_resource_id=$(az vm show -g $my_resource_group -n $my_windows_vm --query "id" -o tsv)
+
+                default_config=$(az vm diagnostics get-default-config  --is-windows-os \\
+                    | sed "s#__DIAGNOSTIC_STORAGE_ACCOUNT__#$my_diagnostic_storage_account#g" \\
+                    | sed "s#__VM_OR_VMSS_RESOURCE_ID__#$my_vm_resource_id#g")
+
+                # Please use the same options, the WAD diagnostic extension has strict
+                # expectations of the sas token's format. Set the expiry as desired.
+                storage_sastoken=$(az storage account generate-sas \\
+                    --account-name $my_diagnostic_storage_account --expiry 2037-12-31T23:59:00Z \\
+                    --permissions acuw --resource-types co --services bt --https-only --output tsv)
+
+                protected_settings="{'storageAccountName': '$my_diagnostic_storage_account', \\
+                    'storageAccountSasToken': '$storage_sastoken'}"
+
+                # # Alternatively, if the WAD extension has issues parsing the sas token,
+                # # one can use a storage account key.
+                # storage_account_key=$(az storage account keys list --account-name tosinstorage1win \\
+                #   --query [0].value -o tsv)
+                # protected_settings="{'storageAccountName': '$my_diagnostic_storage_account', \\
+                #   'storageAccountKey': '$storage_account_key'}"
+
+                az vm diagnostics set --settings "$default_config" \\
+                    --protected-settings "$protected_settings" \\
+                    --resource-group $my_resource_group --vm-name $my_windows_vm
 """
 
 disk_long_summary = """
-        Just like any other computer, virtual machines in Azure use disks as a place to store an operating system, applications, and data.
+        Azure Virtual Machines use disks as a place to store an operating system, applications, and data.
         All Azure virtual machines have at least two disks: An operating system disk, and a temporary disk.
         The operating system disk is created from an image, and both the operating system disk and the image are actually virtual hard disks (VHDs)
         stored in an Azure storage account. Virtual machines also can have one or more data disks, that are also stored as VHDs.
 
-        Operating System Disk
-        Every virtual machine has one attached operating system disk. It's registered as a SATA drive and is labeled /dev/sda by default.
-        This disk has a maximum capacity of 1023 gigabytes (GB).
 
-        Temporary disk
-        The temporary disk is automatically created for you. On Linux virtual machines, the disk is typically /dev/sdb and is formatted and
-        mounted to /mnt/resource by the Azure Linux Agent. The size of the temporary disk varies, based on the size of the virtual machine.
+        Azure Managed and Unmanaged Data Disks have a maximum size of 4095 GB (with the exception of larger disks in preview). Azure Unmanaged Disks also have a maximum capacity of 4095 GB.
 
-        Data disk
-        A data disk is a VHD that's attached to a virtual machine to store application data, or other data you need to keep. Data disks are
-        registered as SCSI drives and are labeled by the creator. Each data disk has a maximum capacity of 1023 GB. The size of the virtual
-        machine determines how many data disks can be attached and the type of storage that can be used to host the disks.
+
+        For more information, see:
+
+        - Azure Disks - https://docs.microsoft.com/en-us/azure/virtual-machines/linux/about-disks-and-vhds and https://docs.microsoft.com/en-us/azure/virtual-machines/windows/about-disks-and-vhds.
+
+        - Larger Managed Disks in Public Preview - https://azure.microsoft.com/en-us/blog/introducing-the-public-preview-of-larger-managed-disks-sizes/
+
+        - Ultra SSD Managed Disks in Public Preview - https://docs.microsoft.com/en-us/azure/virtual-machines/windows/disks-ultra-ssd
+
 """
 
 helps['vm disk'] = """
@@ -610,7 +643,7 @@ helps['vm unmanaged-disk attach'] = """
     long-summary: This allows for the preservation of data, even if the VM is reprovisioned due to maintenance or resizing.
     examples:
         - name: Attach a new default sized (1023 GB) unmanaged data disk to a VM.
-          text: az vm unmanaged-disk attach -g MyResourceGroup --vm-name MyVm
+          text: az vm unmanaged-disk attach -g MyResourceGroup --vm-name MyVm --new
         - name: Attach an existing data disk to a VM as unmanaged.
           text: >
             az vm unmanaged-disk attach -g MyResourceGroup --vm-name MyVm \\
@@ -632,10 +665,10 @@ helps['vm unmanaged-disk list'] = """
     examples:
         - name: List the unmanaged disks attached to a VM.
           text: az vm unmanaged-disk list -g MyResourceGroup --vm-name MyVm
-        - name: List unmanaged disks with IDs containing the string "data_disk".
+        - name: List unmanaged disks with names containing the string "data_disk".
           text: >
-            az vm unmanaged-disk list --ids \\
-                $(az resource list --query "[?contains(name, 'data_disk')].id" -o tsv)
+            az vm unmanaged-disk list -g MyResourceGroup --vm-name MyVm \\
+                --query "[?contains(name, 'data_disk')]" --output table
 """
 
 helps['vm disk detach'] = """
@@ -644,7 +677,7 @@ helps['vm disk detach'] = """
     examples:
         - name: Detach a data disk from a VM.
           text: >
-            az vm disk detach -g MyResourceGroup --vm-name MyVm -n disk_name
+            az vm disk detach -g MyResourceGroup --vm-name MyVm --name disk_name
 """
 
 helps['vm disk attach'] = """
@@ -653,7 +686,7 @@ helps['vm disk attach'] = """
     long-summary: This allows for the preservation of data, even if the VM is reprovisioned due to maintenance or resizing.
     examples:
         - name: Attach a new default sized (1023 GB) managed data disk to a VM.
-          text: az vm disk attach -g MyResourceGroup --vm-name MyVm --disk disk_name --new
+          text: az vm disk attach -g MyResourceGroup --vm-name MyVm --name disk_name --new
 """
 
 helps['vm encryption'] = """
@@ -970,6 +1003,17 @@ helps['vmss extension set'] = """
     - name: --name -n
       populator-commands:
       - az vm extension image list
+    examples:
+        - name: >
+                Set an extension which depends on two previously set extensions. That is, When a VMSS instance is
+                created or reimaged, the customScript extension will be provisioned only after all extensions that
+                it depends on have been provisioned. The extension need not depend on the other extensions for
+                pre-requisite configurations.
+          text: >
+                az vmss extension set --vmss-name my-vmss --name customScript --resource-group my-group \\
+                    --version 2.0 --publisher Microsoft.Azure.Extensions \\
+                    --provision-after-extensions NetworkWatcherAgentLinux VMAccessForLinux  \\
+                    --settings '{\"commandToExecute\": \"echo testing\"}'
 """
 
 helps['vmss extension show'] = """
@@ -1324,7 +1368,9 @@ helps['vmss identity show'] = """
 helps['disk'] = """
     type: group
     short-summary: Manage Azure Managed Disks.
-"""
+    long-summary: >
+{0}
+""".format(disk_long_summary)
 
 helps['image'] = """
     type: group
@@ -1504,22 +1550,29 @@ helps['sig image-version create'] = """
     short-summary: creat a new image version
     long-summary : this operation might take a long time depending on the replicate region number. Use "--no-wait" is advised.
     examples:
-        - name: add a new image version
+        - name: Add a new image version
           text: |
             az sig image-version create -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --managed-image /subscriptions/00000000-0000-0000-0000-00000000xxxx/resourceGroups/imageGroups/providers/images/MyManagedImage
-        - name: add a new image version and don't wait on it. Later you can invoke "az sig image-version wait" command when ready to create a vm from the gallery image version
+        - name: Add a new image version replicated across multiple regions with different replication counts each. Eastus2 will have it's replica count set to the default replica count.
           text: |
-            az sig image-version create --no-wait -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --managed-image imageInTheSameResourceGroup
+                az sig image-version create -g MyResourceGroup --gallery-name MyGallery \\
+                --gallery-image-definition MyImage --gallery-image-version 1.0.0 \\
+                --managed-image image-name --target-regions eastus2 ukwest=3 southindia=2
+        - name: Add a new image version and don't wait on it. Later you can invoke "az sig image-version wait" command when ready to create a vm from the gallery image version
+          text: |
+            az sig image-version create --no-wait -g MyResourceGroup --gallery-name MyGallery \\
+            --gallery-image-definition MyImage --gallery-image-version 1.0.0 \\
+            --managed-image imageInTheSameResourceGroup
 """
 
 helps['sig image-version update'] = """
     type: command
     short-summary: update a share image version
     examples:
-        - name: replicate to a new set of regions
+        - name: Replicate to a new set of regions
           text: |
             az sig image-version update -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --target-regions westcentralus=2 eastus2
-        - name: replicate to one more region
+        - name: Replicate to one more region
           text: |
             az sig image-version update -g MyResourceGroup --gallery-name MyGallery --gallery-image-definition MyImage --gallery-image-version 1.0.0 --add publishingProfile.targetRegions name=westcentralus